Changed Functional/Transducer folder topology
