+++ src/Functional/Transducer/index.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * transducer functions
+ *
+ * @package bingo-functional
+ * @author Lochemem Bruno Michael
+ * @license Apache-2.0
+ */
+
+require_once __DIR__ . '/Filter.php';
+require_once __DIR__ . '/Map.php';
+require_once __DIR__ . '/Reject.php';
